Version 1.3
Resets connections to Splunk server if they are ever closed. This allows
LARGE amounts of data to be sent.

But sending millions of events causes a large spike of memory usage.
This is because of the ParseXML function not streaming in the data. Next
step is to look into better methods to doing this.